Lending Commit: Modified Controllers

+++ Lending/ApiControllers/ApiApplicantApplianceOwnedController.cs
@@ -40,15 +40,30 @@ namespace Lending.ApiControllers
         {
             try
             {
-                Data.mstApplicantApplianceOwned newApplianceOwned = new Data.mstApplicantApplianceOwned();
-                newApplianceOwned.ApplicantId = applianceOwned.ApplicantId;
-                newApplianceOwned.ApplianceBrand = applianceOwned.ApplianceBrand;
-                newApplianceOwned.PresentValue = applianceOwned.PresentValue;
+                var applicants = from d in db.mstApplicants where d.Id == Convert.ToInt32(applianceOwned.ApplicantId) select d;
+                if (applicants.Any())
+                {
+                    if (!applicants.FirstOrDefault().IsLocked)
+                    {
+                        Data.mstApplicantApplianceOwned newApplianceOwned = new Data.mstApplicantApplianceOwned();
+                        newApplianceOwned.ApplicantId = applianceOwned.ApplicantId;
+                        newApplianceOwned.ApplianceBrand = applianceOwned.ApplianceBrand;
+                        newApplianceOwned.PresentValue = applianceOwned.PresentValue;
 
-                db.mstApplicantApplianceOwneds.InsertOnSubmit(newApplianceOwned);
-                db.SubmitChanges();
+                        db.mstApplicantApplianceOwneds.InsertOnSubmit(newApplianceOwned);
+                        db.SubmitChanges();
 
-                return Request.CreateResponse(HttpStatusCode.OK);
+                        return Request.CreateResponse(HttpStatusCode.OK);
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
+                }
+                else
+                {
+                    return Request.CreateResponse(HttpStatusCode.NotFound);
+                }
             }
             catch
             {
@@ -64,17 +79,32 @@ namespace Lending.ApiControllers
         {
             try
             {
-                var applianceOwneds = from d in db.mstApplicantApplianceOwneds where d.Id == Convert.ToInt32(id) select d;
-                if (applianceOwneds.Any())
+                var applicants = from d in db.mstApplicants where d.Id == Convert.ToInt32(applianceOwned.ApplicantId) select d;
+                if (applicants.Any())
                 {
-                    var updateApplianceOwned = applianceOwneds.FirstOrDefault();
-                    updateApplianceOwned.ApplicantId = applianceOwned.ApplicantId;
-                    updateApplianceOwned.ApplianceBrand = applianceOwned.ApplianceBrand;
-                    updateApplianceOwned.PresentValue = applianceOwned.PresentValue;
+                    if (!applicants.FirstOrDefault().IsLocked)
+                    {
+                        var applianceOwneds = from d in db.mstApplicantApplianceOwneds where d.Id == Convert.ToInt32(id) select d;
+                        if (applianceOwneds.Any())
+                        {
+                            var updateApplianceOwned = applianceOwneds.FirstOrDefault();
+                            updateApplianceOwned.ApplicantId = applianceOwned.ApplicantId;
+                            updateApplianceOwned.ApplianceBrand = applianceOwned.ApplianceBrand;
+                            updateApplianceOwned.PresentValue = applianceOwned.PresentValue;
 
-                    db.SubmitChanges();
+                            db.SubmitChanges();
 
-                    return Request.CreateResponse(HttpStatusCode.OK);
+                            return Request.CreateResponse(HttpStatusCode.OK);
+                        }
+                        else
+                        {
+                            return Request.CreateResponse(HttpStatusCode.NotFound);
+                        }
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.BadRequest);
+                    }
                 }
                 else
                 {
@@ -98,10 +128,25 @@ namespace Lending.ApiControllers
                 var applianceOwneds = from d in db.mstApplicantApplianceOwneds where d.Id == Convert.ToInt32(id) select d;
                 if (applianceOwneds.Any())
                 {
-                    db.mstApplicantApplianceOwneds.DeleteOnSubmit(applianceOwneds.First());
-                    db.SubmitChanges();
+                    var applicants = from d in db.mstApplicants where d.Id == applianceOwneds.FirstOrDefault().ApplicantId select d;
+                    if (applicants.Any())
+                    {
+                        if (!applicants.FirstOrDefault().IsLocked)
+                        {
+                            db.mstApplicantApplianceOwneds.DeleteOnSubmit(applianceOwneds.First());
+                            db.SubmitChanges();
 
-                    return Request.CreateResponse(HttpStatusCode.OK);
+                            return Request.CreateResponse(HttpStatusCode.OK);
+                        }
+                        else
+                        {
+                            return Request.CreateResponse(HttpStatusCode.BadRequest);
+                        }
+                    }
+                    else
+                    {
+                        return Request.CreateResponse(HttpStatusCode.NotFound);
+                    }
                 }
                 else
                 {
